fix: allow consecutive `insert.text` events
The `validateSelection` logic was causing this to fail in a confusing way since
it would cause the editor to be selected at `[0, 0]` after the second
`insert.text` event. Since PTE has become a lot more stable (and better at
cathing errors thrown by Slate), this logic might not be needed anymore
(it's hard to tell without any tests *verifying* that the logic *is* needed).
Therefore, it has been removed in order to allow consecutive `insert.text`
events.