fix: harden outgoing patch logic surrounding clearing the editor value
The placeholder block is now only inserted declaratively through normalization
if the editor has been made empty. The only exception to this rule is when an
`'insert.block'` event is performed after the editor is cleared. This could
happen in a Behavior that stacks events so they happen without normalization
running in between. For now, `'insert.block'` always expects at least a
placeholder, and therefore one is added if needed.

If the editor was cleared by remote changes, the placeholder is inserted
without emitting any patches telling that it happened.

A placeholder is a text block with no `listItem`, no inline objects and no text.
It can have a `style` but in that case it has to be the default style (usually
`'normal'`.

An editor containing only a placeholder is considered empty, unless the
placeholder is equal to the `initialValue` of the editor. This has to do with
the fact that the editor automatically emits `setIfMissing` and `insert`
patches to insert the placeholder in the foreign value, when the editor goes
from empty to non-empty. But if the `initialValue` resembles a placeholder,
then this would cause an extra placeholder insertion in the foreign value.

To make the empty editor check more robust, the editor is deemed non-empty if
the placeholder has any custom props. This, `set`ing custom props on the
placeholder will force an `insert` patch to be emitted, updating the foreign
value to include the placeholder.

Lots of new test cases have been added to ensure correct placeholder logic and
outgoing patches, including test cases that mimic a "foreign value" updated by
the editor's outgoing patches.