fix: replace WeakMaps with direct editor properties
All WeakMaps have been migrated to properties on the `editor` object. This
removes indirection, simplifies the code and makes debugging easier.

WeakMaps are useful in Slatewhere they need to attach state to objects they
don't control. But since we own the `PortableTextSlateEditor` interface and
control the editor's entire lifecycle, we can simply add properties directly
to the editor object.