fix: make unique key checks faster
Instead of getting the keys of all blocks in the editor every time a block is
inserted (O(N)), we check if the `blockIndexMap` has the key (O(1)). For span
and inline object insertion we still iterate through the parent text block's
children, but these arrays are usually fairly small.